Efficient DOM-non-update event handling (#40)
Default implementation of concur-replica is to update DOM after handling a single event. This is impractical if we handle very frequent event firing like onMouseMove since DOM update is very expensive as computing diff and sending diff patch to websocket happen. In this PR, I modified the source of concur-replica and replica (that part is embedded to this repo) to support non-updating state of DOM, which only generates new events without DOM change.

* onMouseMove test
* retrieve mouse coordinates
* embed lastUpdated into UIState
* more clear separation of IO action insertion: unsafeBlockingIO vs liftIO
* limit mouse move events by time difference.
* modify concur-* upstream with onlyEvent
* change HTML -> IHTML
* internalize concur-replica runner and handler
* IHTML is now different from HTML
* extend IHTML ~ Either HTML HTML, where Left represents no update VDOM.
* handle Left case w/o diff update.
* move onMouseMove -> ConcurReplica.DOM.Events
* block / unblock DOM update!
* separate out serverUpdatedTime and uiUpdatedTime
* added comments