Introducing DriverId (#62)
Previously, ModuleName is used as a key for inspecting on each module unit compilation. 
Unfortunately, the ModuleName info was not available at the start time of module compilation 
(this is the limitation of driverPlugin, though. GHC compileOne function invocation already has
module summary info when initializing the plugin for the compilation unit), so the previous 
method was ad hoc to keep the module start time information until the module name info is 
available and send together.
Now, DriverId is assigned for each instance of driver, and send start time right at the start time 
and then later module name information is separately sent. This simplifies the internal logic and 
enable inspection before the module name arrives.

* PluginSession
* introduce module id issuer
* extend timing message with driver id and optional module name
* newtype DriverId
* timing message only use driver id. driver id ~ ModuleName connection from separate message
* rendering update with DriverId
* rename: isInProgress -> isTimeInTimerRange, isCompiled -> isModuleCompilationDone
* introduce BiMap
* use KeyMap and BiKeyMap. no more bare Int for DriverId