Race condition in initialization correctly handled (#98)
Read session and overwrite on the session should be done in a single atomic STM operation.
Unfortunately, as a consequence, the necessary IO action should be done outside beforehand, which means the action will be done multiple times per every driver start. This is due to the limited plugin point (as we cannot put a compatible plugin in the beginning of GHC session. There exists frontend plugin, but that's not compatible with cabal, which uses hard-coded ghc --make command yet) The IO operation is nothing but timestamping and reading config, so I put up with a little redundancy here.

* Race condition in initialization correctly handled
* made comment.